harmonive solve_network across workflow
