+++ config.default.yaml
@@ -629,15 +629,7 @@ solving:
     track_iterations: false
     min_iterations: 4
     max_iterations: 6
-    keep_shadowprices:
-    - Bus
-    - Line
-    - Link
-    - Transformer
-    - GlobalConstraint
-    - Generator
-    - Store
-    - StorageUnit
+    seed: 123
 
   solver:
     name: gurobi
