+++ rules/solve_overnight.smk
@@ -35,4 +35,4 @@ rule solve_sector_network:
     conda:
         "../envs/environment.yaml"
     script:
-        "../scripts/solve_sector_network.py"
+        "../scripts/solve_network.py"
